stage1_dataset/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fly-olf-stage1"
version = "0.1.0"
description = "Stage 1 pipeline for fly olfactory learning: trial segmentation and feature extraction"
requires-python = ">=3.9"
dependencies = [
    "pandas>=1.5",
    "numpy>=1.23",
    "pyarrow>=10.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "matplotlib>=3.5",
    "seaborn>=0.12",
]

[project.scripts]
fly-olf-stage1 = "fly_olf.stage1.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
